refactor: HandlerMappings 클래스 제거
